requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.3.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.8
pandas==1.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
